cilium-dbg: Fix map not found bugs on IPv6-only clusters
On IPv6-only clusters, several of our cilium-dbg bpf commands fail
because we try to open IPv4 maps and they don't exist. This commit fixes
it by extending (and renaming) the getIpv6EnableStatus helper to detect
both the IPv4 and IPv6 statuses in the agent. This information is then
passed to the map creation functions.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>